refactor all logic to the use form hook
