+++ app/page.tsx
@@ -1,63 +1,17 @@
 'use client'
-import {useReducer, useState} from "react";
 import Select from "@/components/Select";
 import TextInput from "@/components/TextInput";
 import CheckBox from "@/components/CheckBox";
 import SummaryTable from "@/components/SummaryTable";
-import {act} from "react-dom/test-utils";
+import useForm from "@/hooks/useForm";
 
 export default function Home() {
-  const initialState = {
-    inp1: '',
-    inp2: '',
-    sel1: '',
-    cb1: false,
-    cb2: false,
-    cb3: false
-  }
-  const reducer = (state, action) => {
-    switch (action.type) {
-      case 'SET_TEXT': return {
-        ...state,
-        [action.payload.name]: action.payload.value
-      }
-
-      case 'SET_SELECT': return {
-        ...state,
-        [action.payload.name]: action.payload.value
-      }
-
-      case 'TOGGLE_CHECKBOX': return {
-        ...state,
-        [action.payload.name]: !state[action.payload.name]
-      }
-
-      default: return state
-    }
-  }
-
-  const [state, dispatch] = useReducer(reducer, initialState)
-
-  const handleInputChange = ({ target }) => {
-    dispatch({
-      type: 'SET_TEXT',
-      payload: { name: target.name, value: target.value }
-    })
-  }
-
-  const handleSelectChange = ({ target }) => {
-    dispatch({
-      type: 'SET_SELECT',
-      payload: { name: target.name, value: target.value }
-    })
-  }
-
-  const handleCheckboxClick = ({ target }) => {
-    dispatch({
-      type: 'TOGGLE_CHECKBOX',
-      payload: { name: target.name }
-    })
-  }
+  const {
+    state,
+    handleInputChange,
+    handleSelectChange,
+    handleCheckboxClick
+  } = useForm()
 
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
